Fixed git check

+++ install_git_homedir.sh
@@ -4,6 +4,7 @@
 tmpdir="/tmp/homedirtmp"
 backupdir="$HOME/homedir_backup"
 installdir="$HOME"
+gitinstalled="unknown"
 
 # Check if git is installed
 if git --version; then
@@ -23,7 +24,6 @@ fi
 # check if there is a git repo in $installdir
 if [ -d $installdir/.git  ]; then
     echo "There is a git repo in $installdir, aborting."
-    exit 1
 fi
 
 
@@ -38,17 +38,18 @@ else
 fi
 
 # Continue?
-echo "Do you want to replace these files(y/n):"
-echo "$files"
-read replacefiles
-
-if [ $replacefiles == "y" ]; then
-# Backup files that will be overwritten by homedir.git
-    mkdir $backupdir
-    for file in $files; do
-        echo "Backing up this file: $file"
-        cp $file $backupdir && echo -e "\nFile have been backed up to $backupdir."
-    done
+if [ $gitinstalled = "true" ]; then
+    echo "Do you want to replace these files(y/n):"
+    echo "$files"
+    read replacefiles
+    if [ $replacefiles == "y" ]; then
+ # Backup files that will be overwritten by homedir.git
+        mkdir $backupdir
+        for file in $files; do
+            echo "Backing up this file: $file"
+            cp $file $backupdir && echo -e "\nFile have been backed up to $backupdir."
+        done
 # replace the files
-    cp -r $tmpdir/{,.??}* $installdir && echo "Files have been replaced."
+        cp -r $tmpdir/{,.??}* $installdir && echo "Files have been replaced."
+    fi
 fi
